aafield/requirements.txt
Django==2.2.6
pytz==2019.3
whitenoise==4.1.4
gunicorn==20.0.0
psycopg2==2.7.5
dj-config-url==0.1.1
dj-database-url==0.5.0
django-crispy-forms==1.7.2
django-filter==2.2.0
django_tables2==2.2.1
Pillow==7.0.0
braintree==4.0.0
boto3
django-storages==1.7.1